# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Janus API
  version: "0.1.0"
  description: >
    Self-hosted, single-tenant secrets manager. All routes are under /v1.
    Authenticate with a service token (Authorization: Bearer janus_svc_…) or a
    UI session cookie. Secret values are never returned in list/metadata views.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Service token (janus_svc_… / janus_user_…).
    sessionCookie:
      type: apiKey
      in: cookie
      name: janus_session
  schemas:
...
    Project:
      type: object
      properties:
        id: { type: string, format: uuid }
        slug: { type: string, example: acme-web }
        name: { type: string, example: Acme Web }
        created_at: { type: string, format: date-time }
        last_activity_at:
          type: string
          format: date-time
          nullable: true
          description: Most recent config-version timestamp across the project's configs; null if none.
...
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1 }
      description: Opt-in cursor pagination page size. Omit for the full list.
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string }
  responses:
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
security:
  - bearerAuth: []
  - sessionCookie: []
...
paths:
...
  /v1/projects:
    get:
      tags: [projects]
      summary: List projects
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        "200":
          description: Projects (metadata only).
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items: { $ref: '#/components/schemas/Project' }
                  next_cursor: { type: [string, "null"] }
        "401": { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug]
              properties:
                slug: { type: string, example: acme-web }
                name: { type: string, example: Acme Web }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        "400": { description: Validation error. }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /v1/projects/{pid}:
    parameters:
      - { name: pid, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [projects]
      summary: Get a project
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [projects]
      summary: Rename a project (name only; slug is immutable)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        "200":
          description: Renamed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        "400": { description: Validation error. }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [projects]
      summary: Soft-delete (or, with ?destroy=true, hard-destroy) a project
      parameters:
        - name: destroy
          in: query
          required: false
          schema: { type: boolean, default: false }
          description: When true, permanently destroys instead of soft-deleting.
      responses:
        "204": { description: Deleted. }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
  /v1/projects/{pid}/restore:
    parameters:
      - { name: pid, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [projects, trash]
      summary: Restore a soft-deleted project
      responses:
        "200":
          description: Restored.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
  /v1/projects/{pid}/members:
    parameters:
      - { name: pid, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [members]
      summary: List project members
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        "200":
          description: Project role bindings.
          content:
            application/json:
              schema:
                type: object
                properties:
                  members:
                    type: array
                    items:
                      type: object
                      properties:
                        user_id: { type: string }
                        role: { type: string, enum: [viewer, developer, admin, owner] }
                  next_cursor: { type: [string, "null"] }
        "403": { $ref: '#/components/responses/Forbidden' }
  /v1/projects/{pid}/members/{uid}:
    parameters:
      - { name: pid, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: uid, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [members]
      summary: Grant or update a project member's role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, enum: [viewer, developer, admin, owner] }
      responses:
        "204": { description: Granted. }
        "400": { description: Validation error. }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [members]
      summary: Revoke a project member's role binding
      responses:
        "204": { description: Revoked. }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
  /v1/projects/{pid}/pipeline:
    parameters:
      - { name: pid, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [pipeline]
      summary: Get the project's promotion pipeline (ordered environment ids)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  environment_ids:
                    type: array
                    items: { type: string, format: uuid }
        "403": { $ref: '#/components/responses/Forbidden' }
    put:
      tags: [pipeline]
      summary: Set the project's promotion pipeline order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                environment_ids:
                  type: array
                  items: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  environment_ids:
                    type: array
                    items: { type: string, format: uuid }
        "400": { description: Validation error. }
        "403": { $ref: '#/components/responses/Forbidden' }
  /v1/projects/{pid}/kek:
    parameters:
      - { name: pid, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [kek]
      summary: Get project KEK status (owner-only, value-free)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  current_version: { type: integer }
                  pending:
                    type: array
                    items:
                      type: object
                      properties:
                        version: { type: integer }
                        dek_count: { type: integer }
        "403": { $ref: '#/components/responses/Forbidden' }
  /v1/projects/{pid}/kek/rotate:
    parameters:
      - { name: pid, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [kek]
      summary: Rotate the project KEK to a fresh version (owner-only)
      responses:
        "200":
          description: Rotated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  kek_version: { type: integer }
        "403": { $ref: '#/components/responses/Forbidden' }
        "503": { description: Server is sealed. }
  /v1/projects/{pid}/kek/rewrap:
    parameters:
      - { name: pid, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [kek]
      summary: Re-wrap DEKs still under a superseded KEK version (owner-only, value-free sweep)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  rewrapped: { type: integer }
                  retired_versions:
                    type: array
                    items: { type: integer }
                  remaining: { type: integer }
        "403": { $ref: '#/components/responses/Forbidden' }
        "503": { description: Server is sealed. }
...
  /v1/promote/requests/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [promote]
      summary: Get a promotion request (with a value-free diff)
      responses:
        "200":
          description: >
            Request detail plus a value-free diff (added/changed/removed key
            NAMES only) when the target config already exists.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PromotionRequest' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
...
  /v1/promote/requests/{id}/reject:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [promote]
      summary: Reject a promotion request (requires secret:promote on the target; four-eyes)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                note: { type: string }
      responses:
        "200":
          description: Rejected.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: rejected }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
        "409": { description: Request is no longer pending. }
...
  /v1/auth/logout:
    post:
      tags: [auth]
      summary: Log out and clear the session cookie
      responses:
        "204": { description: Logged out. }
        "401": { $ref: '#/components/responses/Unauthorized' }
...